control page has now 2 toggles for modes
