+++ data/settings.html
@@ -7,7 +7,18 @@
   <link rel="stylesheet" type="text/css" href="style.css">
 </head>
 <body>
-<div class="topnav">
+  <div class="hamburger-menu">
+    <input id="menu__toggle" type="checkbox" />
+    <label class="menu__btn" for="menu__toggle">
+      <span></span>
+    </label>
+
+    <ul class="menu__box">
+      <li><a class="menu__item" href="/control">Home</a></li>
+			<li><a class="menu__item" href="/settings">Network settings</a></li>
+			<li><a class="menu__item" href="/webserial">Webserial</a></li>
+    </ul>
+  </div>
   <h1>Device Settings</h1>
 </div>
 <div class="content">
