+++ include/modules/control.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "Arduino.h"
+#include "main.h"
+
+void control_init();
