+++ platformio.ini
@@ -17,8 +17,8 @@ framework = arduino
 extra_scripts = pre:lib/ProjectUtils-esp32/copy_lib_data.py
 monitor_speed = 115200
 monitor_filters = esp32_exception_decoder
-#upload_protocol = espota
-#upload_port = 192.168.0.50
+upload_protocol = espota
+upload_port = 192.168.0.50
 lib_deps = 
     me-no-dev/AsyncTCP @ ^1.1.1
     https://github.com/me-no-dev/ESPAsyncWebServer.git
